+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 ## Unreleased 0.1.0
+
+### Lists
+
+- `/lists` endpoint added as `Iterable::Lists.new.all`